refactor(decorator-metadata): extract findClassAtPosition to shared ast-utils module
Verified: precommit-verify ✓ (d936df428579e33e)
